+++ src/main/kotlin/com/arkhamusserver/arkhamus/view/dto/netty/response/parts/AbilityOfUserResponse.kt
@@ -4,5 +4,5 @@ data class AbilityOfUserResponse(
     val abilityId: Int? = null,
     val maxCooldown: Long? = null,
     val cooldown: Long = 0,
-    val charges: Long? = null,
+    val charges: Int? = null,
 )
